Also verify whether instructions are allowed to change layouts.
If a test provides a instruction_can_change_layout_func, we should also use it
for the HloVerifier created inside the VerifiedHloModule.
Doing this exposed a bug in algebraic_simplifier where a simplification didn't
check whether it is running in layout sensitive mode (since it doesn't handle
layouts in any way, it should not be used in layout sensitive mode).
Also adjust HloVerifier to be not as strict regarding tiling and memory space
to make a test pass.

PiperOrigin-RevId: 617784203