[xla][gpu] Preserve a frontend attribute for runtime optimization when decomposing a CollectivePermute with a cycle.
The frontend attribute annotates a range of runtime execution instances where
each source-target pair is communicating data that affects the outcome of the
application. As such, when we split the source-target pairs for the decomposed
CollectivePermute, we also need to split the corresponding bounds represented
by the frontend attribute.

Modify two existing tests.

PiperOrigin-RevId: 617846437