[xla:tpu] Allow host offloading ops to pass hlo verification when changing layout.
Host offloading ops (dynamic slice, copy, and dynamic update slice) move tensors between device memory and host memory space. This can cause the hlo verifier to fail as it interprets this layout to be unequal. This change allows for memory space to be ignored when comparing layouts for these ops.

PiperOrigin-RevId: 618677446