[xla:gpu] Use fusion shape instead of hero shape when emit AddressComputationThunk
Supporting DUS in AddressComputationFusion means the fusion shape might be different from the hero custom call shape. Sometimes not the whole tuple returned by the hero custom call op is used, but only the output (e.g. the workspace is ignored for GEMM custom call). As a result, we should emit AddressComputationThunk based on the shape of the fusion instruction and not the one from hero custom call op

PiperOrigin-RevId: 618872694